支持会话消息的增删查

+++ src/reception/database/schema.rs
@@ -39,7 +39,7 @@ diesel::table! {
 
 diesel::table! {
     messages (id) {
-        id -> BigInt,
+        id -> Integer,
         conversation_id -> Integer,
         created_at -> Timestamp,
         content -> Text,
